feat: split Context to ActionContext & StateContext to reduce rerender
